avances de los flujos seraparados de los flujos de usuario

+++ src/app/admin/pages/usuarios/usuarios.component.ts
@@ -58,7 +58,7 @@ export class UsuariosComponent {
        '¿Estás seguro?',
        'Sí, desactivar',
        'Cancelar'
-    ).then((confirmado) => {
+    ).then((confirmado: boolean) => {
        if (confirmado) {
           this.usuarioService.eliminarUsuario(idUsuario).subscribe(
              () => {
